// ani/src/app/models/Select.ts
export interface Sort {
  value: string;
  viewValue: string;
}

export interface Type {
  value: string;
  viewValue: string;
}

export interface Status {
  value: string;
  viewValue: string;
}

export interface Rating {
  value: string;
  viewValue: string;
}

export interface Genre {
  value: string;
  viewValue: string;
}